Refactor package structure of test code files
